+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix survey completed field [#69](https://github.com/rokwire/surveys-building-block/issues/69)
+- 
 ## [1.10.0] - 2024-007-30
 ### Added 
 - Add admin GET survey responses API [#66](https://github.com/rokwire/surveys-building-block/issues/66)